helper.sh: only remove downloaded archives if deleteSource=1

+++ build/media-suite_helper.sh
@@ -138,15 +138,15 @@ do_wget() {
         case $archive_type in
         zip)
             unzip "$archive"
-            rm "$archive"
+            [[ $deleteSource = "y" ]] && rm "$archive"
             ;;
         7z)
             7z x -o"$dirName" "$archive"
-            rm "$archive"
+            [[ $deleteSource = "y" ]] && rm "$archive"
             ;;
         tar*)
             tar -xaf "$archive"
-            rm "$archive"
+            [[ $deleteSource = "y" ]] && rm "$archive"
             cd "$dirName"
             ;;
         esac
